+++ src/main/java/com/blocklegend001/onlyhammersandexcavators/datagen/ModRecipeProvider.java
@@ -34,11 +34,12 @@ public class ModRecipeProvider extends FabricRecipeProvider {
                 .offerTo(exporter);
 
         ShapedRecipeJsonBuilder.create(RecipeCategory.TOOLS, ModItems.STONE_HAMMER)
-                .pattern("BBB")
+                .pattern("BCB")
                 .pattern(" S ")
                 .pattern(" S ")
                 .input('S', Items.STICK)
-                .input('B', ItemTags.STONE_CRAFTING_MATERIALS)
+                .input('B', Blocks.STONE)
+                .input('C', Blocks.COBBLESTONE)
                 .criterion(hasItem(Items.STONE), conditionsFromItem(Items.COBBLESTONE))
                 .offerTo(exporter);